+++ src/handlers/users.rs
@@ -20,11 +20,9 @@ fn list(client: &PgPool) -> impl Filter<Extract = impl Reply, Error = Rejection>
         .fetch_all(&client)
         .await
         {
-            Err(_) => Ok(ResponseBody::<Error>::new_error(Error::new_str(
-                0,
-                "Cannot get the users.",
-            ))
-            .to_reply()),
+            Err(_) => {
+                Ok(ResponseBody::new(400, Error::new_str(0, "Cannot get the users.")).to_reply())
+            }
             Ok(rows) => {
                 let users: Vec<UserModel> = rows
                     .iter()
@@ -55,19 +53,19 @@ fn find(client: &PgPool) -> impl Filter<Extract = impl Reply, Error = Rejection>
 
     async fn handler(query: FindQuery, client: PgPool) -> Result<impl Reply, Infallible> {
         if query.id.is_some() && query.username.is_some() {
-            Ok(ResponseBody::<Error>::new_error(Error::new_str(
-                0,
-                "Cannot find a user by id and username at the same time.",
-            ))
+            Ok(ResponseBody::new(
+                400,
+                Error::new_str(0, "Cannot find a user by id and username at the same time."),
+            )
             .to_reply())
         } else if let Some(id) = query.id {
             match UserModel::from_id(&client, &id, false).await {
                 Err(e) => {
                     eprintln!("User by id: Error: {:?}", e);
-                    Ok(ResponseBody::<Error>::new_error(Error::new(
-                        1,
-                        format!("Cannot find the user #{}", id),
-                    ))
+                    Ok(ResponseBody::new(
+                        400,
+                        Error::new(1, format!("Cannot find the user #{}", id)),
+                    )
                     .to_reply())
                 }
                 Ok(user) => Ok(ResponseBody::new_success(user).to_reply()),
@@ -76,16 +74,16 @@ fn find(client: &PgPool) -> impl Filter<Extract = impl Reply, Error = Rejection>
             match UserModel::from_username(&client, &username, false).await {
                 Err(e) => {
                     eprintln!("User by username: Error: {:?}", e);
-                    Ok(ResponseBody::<Error>::new_error(Error::new(
-                        1,
-                        format!("Cannot find the user by its name: {}", username),
-                    ))
+                    Ok(ResponseBody::new(
+                        400,
+                        Error::new(1, format!("Cannot find the user by its name: {}", username)),
+                    )
                     .to_reply())
                 }
                 Ok(user) => Ok(ResponseBody::new_success(user).to_reply()),
             }
         } else {
-            Ok(ResponseBody::<Error>::new_error(Error::new_str(2, "Invalid query.")).to_reply())
+            Ok(ResponseBody::new(400, Error::new_str(2, "Invalid query.")).to_reply())
         }
     }
 
@@ -120,28 +118,26 @@ struct SignUpAndSignInResponse {
 fn signup(client: &PgPool) -> impl Filter<Extract = impl Reply, Error = Rejection> + Clone {
     async fn handler(body: SignUpAndSigInBody, client: PgPool) -> Result<impl Reply, Infallible> {
         if body.username.is_empty() {
-            return Ok(ResponseBody::<Error>::new_error(Error::new_str(
-                0,
-                "You must enter the username.",
-            ))
-            .to_reply());
+            return Ok(
+                ResponseBody::new(400, Error::new_str(0, "You must enter the username."))
+                    .to_reply(),
+            );
         } else if body.username.len() < 4 || body.username.len() > 15 {
-            return Ok(ResponseBody::<Error>::new_error(Error::new_str(
-                1,
-                "The username must be between 4 and 15 characteres.",
-            ))
+            return Ok(ResponseBody::new(
+                400,
+                Error::new_str(1, "The username must be between 4 and 15 characteres."),
+            )
             .to_reply());
         } else if body.password.is_empty() {
-            return Ok(ResponseBody::<Error>::new_error(Error::new_str(
-                2,
-                "You must enter the password.",
-            ))
-            .to_reply());
+            return Ok(
+                ResponseBody::new(400, Error::new_str(2, "You must enter the password."))
+                    .to_reply(),
+            );
         } else if body.password.len() < 8 || body.password.len() > 40 {
-            return Ok(ResponseBody::<Error>::new_error(Error::new_str(
-                3,
-                "The password must be between 8 and 40 characteres.",
-            ))
+            return Ok(ResponseBody::new(
+                400,
+                Error::new_str(3, "The password must be between 8 and 40 characteres."),
+            )
             .to_reply());
         }
 
@@ -152,15 +148,15 @@ fn signup(client: &PgPool) -> impl Filter<Extract = impl Reply, Error = Rejectio
         {
             Err(e) => {
                 eprintln!("SignUp: Error: {:?}", e);
-                Ok(ResponseBody::<Error>::new_error(Error::new_str(5, "Unknown.")).to_reply())
+                Ok(ResponseBody::new(400, Error::new_str(5, "Unknown.")).to_reply())
             }
             Ok(with_username) => {
                 let count: i64 = with_username.get(0);
                 if count > 0 {
-                    return Ok(ResponseBody::<Error>::new_error(Error::new_str(
-                        4,
-                        "The username already exists.",
-                    ))
+                    return Ok(ResponseBody::new(
+                        400,
+                        Error::new_str(4, "The username already exists."),
+                    )
                     .to_reply());
                 }
 
@@ -181,7 +177,7 @@ fn signup(client: &PgPool) -> impl Filter<Extract = impl Reply, Error = Rejectio
                     .await {
                     Err(e) => {
                         eprintln!("Signup: Error: {:?}", e);
-                        Ok(ResponseBody::<Error>::new_error(Error::new_str(5, "Cannot create the user.")).to_reply())
+                        Ok(ResponseBody::new(400, Error::new_str(5, "Cannot create the user.")).to_reply())
                     },
                     Ok(result) => {
                         if result.rows_affected() == 1 {
@@ -191,7 +187,7 @@ fn signup(client: &PgPool) -> impl Filter<Extract = impl Reply, Error = Rejectio
                                 profile_image: user.get_profile_image(),
                             }).to_reply())
                         } else {
-                            Ok(ResponseBody::<Error>::new_error(Error::new_str(5, "Rows not affected.")).to_reply())
+                            Ok(ResponseBody::new(400, Error::new_str(5, "Rows not affected.")).to_reply())
                         }
                     },
                 }
@@ -215,35 +211,31 @@ fn signup(client: &PgPool) -> impl Filter<Extract = impl Reply, Error = Rejectio
 fn signin(client: &PgPool) -> impl Filter<Extract = impl Reply, Error = Rejection> + Clone {
     async fn handler(body: SignUpAndSigInBody, client: PgPool) -> Result<impl Reply, Infallible> {
         if body.username.is_empty() {
-            return Ok(ResponseBody::<Error>::new_error(Error::new_str(
-                0,
-                "You must enter the username.",
-            ))
-            .to_reply());
+            return Ok(
+                ResponseBody::new(400, Error::new_str(0, "You must enter the username."))
+                    .to_reply(),
+            );
         } else if body.password.is_empty() {
-            return Ok(ResponseBody::<Error>::new_error(Error::new_str(
-                1,
-                "You must enter the password.",
-            ))
-            .to_reply());
+            return Ok(
+                ResponseBody::new(400, Error::new_str(1, "You must enter the password."))
+                    .to_reply(),
+            );
         }
 
         match UserModel::from_username(&client, &body.username, true).await {
             Err(e) => {
                 eprintln!("Signin: Error: {:?}", e);
-                Ok(ResponseBody::<Error>::new_error(Error::new_str(
-                    2,
-                    "The username does not exist.",
-                ))
-                .to_reply())
+                Ok(
+                    ResponseBody::new(400, Error::new_str(2, "The username does not exist."))
+                        .to_reply(),
+                )
             }
             Ok(user) => {
                 if !user.verify_password(body.password) {
-                    Ok(ResponseBody::<Error>::new_error(Error::new_str(
-                        3,
-                        "The password is incorrect.",
-                    ))
-                    .to_reply())
+                    Ok(
+                        ResponseBody::new(400, Error::new_str(3, "The password is incorrect."))
+                            .to_reply(),
+                    )
                 } else {
                     Ok(ResponseBody::new_success(SignUpAndSignInResponse {
                         id: user.get_id().to_string(),